split out grouper.py
