+++ pandas/core/groupby/__init__.py
@@ -1,4 +1,4 @@
 from pandas.core.groupby.groupby import GroupBy  # flake8: noqa
 from pandas.core.groupby.generic import (
     SeriesGroupBy, DataFrameGroupBy, PanelGroupBy)  # flake8: noqa
-from pandas.core.groupby.ops import Grouper  # flake8: noqa
+from pandas.core.groupby.grouper import Grouper  # flake8: noqa
