+++ pandas/tests/groupby/aggregate/test_aggregate.py
@@ -10,7 +10,7 @@ import numpy as np
 import pandas as pd
 
 from pandas import concat, DataFrame, Index, MultiIndex, Series
-from pandas.core.groupby.ops import Grouping
+from pandas.core.groupby.grouper import Grouping
 from pandas.core.base import SpecificationError
 from pandas.compat import OrderedDict
 import pandas.util.testing as tm
